# icalrecur.c: Some performance improvements.