Implement a number of performance improvements.

Most notable changes:
* Avoid evaluating limiting expressions in `check_contracting_rules()` if not needed. I.e. we don't need to calculate day of week, day of month and week no if no according restriction is applied.
* Avoid extra iterations over all days in the 'days' bit field in `expand_by_day()` and `filter_bysetpos()`.

## Benchmark
I benchmarked the change by running the test cases from icalrecur_test 1000 times each. The time in icalrecur.c is as follows:

With ICU
|     | w/o ICU | with ICU |
| --- | ---: | ---: |
| before | 1.300 ms | 7.070 ms |
| after | 700 ms | 4.850 ms |

Tested on a 64bit Ubuntu Docker container hosted on a Windows Laptop (Intel Core i7-10750H).